Begin implementing module and classes. Add test.
This change introduces top level methods and variables for the
Mother lode module, the lode and executor abstractions, and a
NilExecutor class. Lode are associations between commands and executors
that take receivers to determine what functionality to run for a given
command. The module keeps track of registered nodes in a set, ensuring
against collision (the same command being associated to two different
nodes).