Add ComposableExecutor. Update Executor.execute.
This change adds a composable executor. Lodes that want to chain
functionality, e.g. commands that accept subcommands, need to implement
ComposableExecutor to call other executors. The executor function has
been updated to use a NamedTuple argument. This ensures executors that
require arguments have a standardized way of accessing them. This may
change to a hash or other structure later on.