Implement Chips.
This change implements a chip based model instead of that of executor.
Chips need to use a mixture of an interface of functions defined in a
module and a super class. This is due to abstract keyword limitations.
One cannot presently define a class method as abstract, but one can
define an abstract interface and use extend to turn it into a set of
required class methods.